Remove circlular imports
